Add files via upload
For our project we decided to make a custom engine assembly known as a short block, based off of the Chevrolet big block 454. This consists of the engine block, crankshaft, connecting rods and pistons, along with several other parts such as nuts and bolts. By making a short block instead of a full engine, this allows us to engineer a high-performance assembly that can be customized by consumers to fit any specifications desired. With a short block, customers have freedom to equip any type of heads, exhaust, intake, fuel delivery, power adders and so on, so the engine can be geared to run on the street or the drag strip.